commit 03/25

+++ DontStarve_WinApi/DontStarve_Client/02_GameObject/Item/Tool/Tool.h
@@ -20,7 +20,7 @@ public:
     Tool(GameObjectID id, const std::wstring& name, const std::wstring& desc, const std::wstring& baseDir, const std::wstring& imageName, int damage, float attackRange);
     virtual ~Tool();
 
-	float GetDamage()      const { return m_damage; }
+	int GetDamage()      const { return m_damage; }
 	float GetAttackRange() const { return m_attackRange; }
 	bool  CanAttack()      const { return m_canAttack; }
 	ToolState GetToolState() const { return m_toolState; }
